scripts: rename test_eval_20250917 -> test_eval
